fixes to template
