+++ .github/workflows/template_unit_test.yml
@@ -1,4 +1,4 @@
-name: end-to-end 🧪 `template.ipynb`
+name: Unit 🧪 `template.ipynb`
 on:
   push
 
@@ -15,5 +15,5 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
-      - name: End-to-end test 
-        run: pytest -vv tests/test_template_notebook.py 
+      - name: Unit Tests with pytest
+        run: pytest -vv tests/test_template_notebook_unittest.py
